implement deterministic fuzz params for TestConfig (renamed TestFile)
