+++ Cargo.toml
@@ -16,6 +16,7 @@ path = "src/bin/cli.rs"
 alloy = { version = "0.3.1", features = ["full"] }
 clap = { version = "4.5.16", features = ["derive"] }
 lazy_static = "1.5.0"
+rand = "0.8.5"
 serde = "1.0.209"
 sqlite = "0.36.1"
 toml = "0.8.19"
