+++ src/spammer/mod.rs
@@ -1,13 +1,19 @@
-pub mod testfile;
+pub mod rand_seed;
+pub mod test_config;
 pub mod univ2;
 
 use crate::Result;
 use alloy::rpc::types::TransactionRequest;
+use rand_seed::RandSeed;
 
 /// Implement SpamTarget for a specific contract to programmatically
 /// generate templates for advanced testing scenarios.
 pub trait SpamTarget {
-    fn get_spam_txs(&self) -> Result<Vec<TransactionRequest>>;
+    fn get_spam_txs(
+        &self,
+        amount: usize,
+        seed: Option<RandSeed>,
+    ) -> Result<Vec<TransactionRequest>>;
 }
 
 /// The main RPC spam controller; sends transactions to the RPC at a given rate.
